✨stay updated

+++ proyectito/src/main/java/com/practica/proyectito/repository/Movimientos_InventarioRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface Movimientos_InventarioRepository extends JpaRepository<Movimientos_Inventario, Long> {
-    Optional<Movimientos_Inventario> findById(int id);
+    Optional<Movimientos_Inventario> findById(Long id);
 }
